Parse message IDs as floats
(imported from commit d9954e4225e95e2931be1d1fe5ecf0b93914b07e)

+++ static/js/rows.js
@@ -36,7 +36,7 @@ exports.last_visible = function () {
 };
 
 exports.id = function (message_row) {
-    return parseInt(message_row.attr('zid'), 10);
+    return parseFloat(message_row.attr('zid'), 10);
 };
 
 var valid_table_names = {
